--- app/src/main/java/com/gurumlab/vocaroutine/data/source/local/UserDataStore.kt
@@ -1,41 +1,0 @@
-package com.gurumlab.vocaroutine.data.source.local
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.emptyPreferences
-import androidx.datastore.preferences.core.stringPreferencesKey
-import androidx.datastore.preferences.preferencesDataStore
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.map
-import java.io.IOException
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class UserDataStore @Inject constructor(@ApplicationContext private val context: Context) {
-
-    private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "dataStore")
-
-    private val uid = stringPreferencesKey("UID")
-
-    val getUid: Flow<String> = context.dataStore.data
-        .catch { exception ->
-            if (exception is IOException) {
-                emit(emptyPreferences())
-            } else {
-                throw exception
-            }
-        }.map { preference ->
-            preference[uid] ?: ""
-        }
-
-    suspend fun setUid(inputUid: String) {
-        context.dataStore.edit {
-            it[uid] = inputUid
-        }
-    }
-}
