+++ app/src/main/java/com/gurumlab/vocaroutine/di/DataStoreModule.kt
@@ -1,0 +1,28 @@
+package com.gurumlab.vocaroutine.di
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.PreferenceDataStoreFactory
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStoreFile
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+private const val dataStore = "dataStore"
+
+@InstallIn(SingletonComponent::class)
+@Module
+object DataStoreModule {
+
+    @Singleton
+    @Provides
+    fun provideDataStore(@ApplicationContext appContext: Context): DataStore<Preferences> {
+        return PreferenceDataStoreFactory.create(
+            produceFile = { appContext.preferencesDataStoreFile(dataStore) }
+        )
+    }
+}
